requirements.txt
asgiref==3.7.2
distro==1.9.0
Django==5.0.2
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tabula-py==2.9.0
typing_extensions==4.10.0
tzdata==2024.1
